um: install a dummy SIGALRM signal handler
The kselftest timers/posix_timers loops on UML. This is because the test
program installs a timer and the userspace does only loop and does no
syscalls.
With installation of a SIGALRM handler the ptrace wait in the
UML kernel is ended though the timer relay signal.

Test case is:
vmlinux mem=500m rootfstype=hostfs ro init=/pathto/timers/posix_timers

Signed-off-by: Thomas Meyer <thomas@m3y3r.de>